style: move external editor configuration to dedicated Editor setup view
